requirements.txt
Flask
Flask-Injector
Flask-RESTful
Flask-Caching
invoke
coverage
gunicorn
waitress
cassandra-driver
